+++ src/bika/lims/browser/analyses/view.py
@@ -23,6 +23,7 @@ from collections import OrderedDict
 from copy import copy
 from copy import deepcopy
 from datetime import datetime
+from datetime import timedelta
 from operator import itemgetter
 
 from bika.lims import api
@@ -38,9 +39,11 @@ from bika.lims.interfaces import IAnalysisRequest
 from bika.lims.interfaces import IFieldIcons
 from bika.lims.interfaces import IReferenceAnalysis
 from bika.lims.interfaces import IRoutineAnalysis
+from bika.lims.interfaces import ISubmitted
 from bika.lims.utils import check_permission
 from bika.lims.utils import format_supsub
 from bika.lims.utils import formatDecimalMark
+from bika.lims.utils import get_fas_ico
 from bika.lims.utils import get_image
 from bika.lims.utils import get_link
 from bika.lims.utils import get_link_for
@@ -790,6 +793,8 @@ class AnalysesView(ListingView):
         self._folder_item_remarks(obj, item)
         # Renders the analysis conditions
         self._folder_item_conditions(obj, item)
+        # Fill beyond holding time icon
+        self._folder_item_beyond_holding_time(obj, item)
 
         return item
 
@@ -1697,6 +1702,51 @@ class AnalysesView(ListingView):
         conditions = "<br/>".join([to_str(cond) for cond in conditions])
         service = item["replace"].get("Service") or item["Service"]
         item["replace"]["Service"] = "<br/>".join([service, conditions])
+
+    def _folder_item_beyond_holding_time(self, analysis_brain, item):
+        """Adds an icon to the item dictionary if no result for the analysis
+        has been submitted although the analytical holding time is due. It also
+        renders the icon if the result capture took place beyond the analytical
+        holding time
+        """
+        analysis = self.get_object(analysis_brain)
+        if not IRoutineAnalysis.providedBy(analysis):
+            return
+
+        # get the maximum holding time of the analysis
+        max_holding_time = analysis.getMaxHoldingTime()
+        if not max_holding_time:
+            return
+
+        # get the datetime from which the analytical holding time is computed
+        start_date = analysis.getDateSampled()
+        start_date = dtime.to_dt(start_date)
+        if not start_date:
+            return
+
+        # calculate the maximum analytical holding date
+        delta = timedelta(minutes=api.to_minutes(**max_holding_time))
+        max_holding_date = dtime.to_ansi(start_date + delta)
+
+        # maybe the result was captured beyond the analytical holding time
+        if ISubmitted.providedBy(analysis):
+            captured = analysis.getResultCaptureDate()
+            captured = dtime.to_ansi(captured)
+            if captured > max_holding_date:
+                msg = _("Result captured beyond the analytical holding time")
+                icon = get_fas_ico("exclamation-triangle",
+                                   css_class="text-danger",
+                                   title=t(msg))
+                self._append_html_element(item, "ResultCaptureDate", icon)
+            return
+
+        # not yet submitted, maybe the analytical holding time is due?
+        now = dtime.to_ansi(datetime.now())
+        if now > max_holding_date:
+            icon = get_fas_ico("exclamation-triangle",
+                               css_class="text-danger",
+                               title=t(_("Analytical holding time is due")))
+            self._append_html_element(item, "ResultCaptureDate", icon)
 
     def is_method_required(self, analysis):
         """Returns whether the render of the selection list with methods is